refactor: Abstractions → Contracts
