--- Assets/YukimaruGames/Terminal/Runtime/Domain/Abstractions.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 10d4814f31470124fbc29002c4e50e96
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
