+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Contracts/Attributes/TerminalCommandAttribute.cs
