+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Contracts/Exceptions/CommandArgumentException.cs
