+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Contracts/Exceptions/CommandFormatException.cs
