+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Contracts/Interfaces/Repositories/ICommandHistory.cs
