+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Contracts/Interfaces/Services/ICommandAutocomplete.cs
