+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Contracts/Interfaces/Services/ICommandDiscoverer.cs
