+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Contracts/Interfaces/Services/ICommandInvoker.cs
