+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Contracts/Interfaces/Services/ICommandLogger.cs
