+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Contracts/Interfaces/Services/ICommandRegistry.cs
