+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Contracts/Models.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 18072eaa7fe80394c910a14f9f460812
+guid: c23d12049e5bcc945b58782503d81bf8
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
