+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Contracts/Models/Entities/CommandLog.cs
