+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Contracts/Models/ValueObjects/CommandArgument.cs
