+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Contracts/Models/ValueObjects/CommandMeta.cs.meta
