+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Contracts/Models/ValueObjects/CommandSpecification.cs
