+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Contracts/YukimaruGames.Terminal.Domain.Contracts.asmdef
@@ -1,6 +1,6 @@
 {
-    "name": "YukimaruGames.Terminal.Domain.Abstractions",
-    "rootNamespace": "YukimaruGames.Terminal.Domain.Abstractions",
+    "name": "YukimaruGames.Terminal.Domain.Contracts",
+    "rootNamespace": "YukimaruGames.Terminal.Domain.Contracts",
     "references": [
         "GUID:37e9bf61fa8119c4c8e79d1f08c439aa"
     ],
